reorganised javascript files
