+++ app/javascript/Posts.jsx
