+++ app/javascript/routes/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Home from "../components/Home";
-import Posts from "../app/features/posts/Posts";
+import Posts from ;
 import Recipe from "../components/Recipe"
 import NewRecipe from "../components/NewRecipe";
 import Login from "../pages/Login";
